+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - [TD-7340] Support for `dynamic_table` template fields
+- [TD-7298] Add event origin tracking (`event_via`) and detailed field changes to concept audit events
 
 ## [7.11.0] 2025-10-13